+++ sdk/test/metrics/exemplar/always_sample_filter_test.cc
@@ -12,8 +12,8 @@ TEST(AlwaysSampleFilter, SampleMeasurement)
   auto filter = opentelemetry::sdk::metrics::ExemplarFilter::GetAlwaysSampleFilter();
   ASSERT_TRUE(
       filter->ShouldSampleMeasurement(1.0, MetricAttributes{}, opentelemetry::context::Context{}));
-  ASSERT_TRUE(
-      filter->ShouldSampleMeasurement(1l, MetricAttributes{}, opentelemetry::context::Context{}));
+  ASSERT_TRUE(filter->ShouldSampleMeasurement((int64_t)1, MetricAttributes{},
+                                              opentelemetry::context::Context{}));
 }
 
 #endif
